There it is! The Mac port. Search for .mac files :)
Yes, although I have/had a flu, I couldn't sit still. I ported 2/3 of
the apps to the Mac platform. What is missing is the third. there is
though some weird error that I can not explain…it just is weird. o.o
Whatsoever, this is a foundament for future updates. Next, I will be
re-writing the chromium package to allow Mac to recognize .desk files
as scripts for nodejs and alike…stay tuned to hear more from the
Mac-phoenix :)