Deskshell.app has own icon, PHP base implemented, not working
Implemented the icon for the Deskshell bundle. I also added a prototype
PHP script to handle the backend, including the entire back; actually a
whole PHP binary and its dependencies. It grew a little in size. I am
yet missing a PHP based and working WebSock client…the ones I found are
broken/ not up to date. I emailed developers, so expect that to change
soon.